backend/requirements.txt
bcrypt==3.1.7
cffi==1.13.2
Click==7.0
DateTime==4.3
dnspython==1.16.0
Flask==1.1.1
Flask-API==2.0
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-PyMongo==2.3.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pkg-resources==0.0.0
pycparser==2.19
PyJWT==1.7.1
pymongo==3.10.0
python-dotenv==0.10.3
pytz==2019.3
six==1.13.0
Werkzeug==0.16.0
zope.interface==4.7.1
